python/requirements.txt
pycryptodomex~=3.20.0